third_party/blink/web_tests/platform/generic/http/tests/devtools/sources/debugger-ui/ui-source-code-expected.txt

Tests UISourceCode class.


Running: testUISourceCode
Content is requested from SourceCodeProvider.
Callback 1 is invoked.
Callback 2 is invoked.
Callback 3 is invoked.
